Default to F24 setup

+++ vagrant-sync/preconfigure.sh
@@ -1,6 +1,17 @@
 #!/bin/bash
 
-sudo yum -y install libselinux-python python
-sudo mkdir -p /root/.ssh
-sudo cp /home/vagrant/.ssh/authorized_keys /root/.ssh/authorized_keys
-sudo chmod -R go-rwx /root/.ssh
+if type dnf; then
+    PACKAGE_MANAGER=dnf
+    PACKAGES="libselinux-python python python2-dnf"
+else
+    PACKAGE_MANAGER=yum
+    PACKAGES="libselinux-python python"
+fi
+
+sudo $PACKAGE_MANAGER -y install $PACKAGES
+
+if ! sudo test -e /root/.ssh/authorized_keys; then
+    sudo mkdir -p /root/.ssh
+    sudo cp /home/vagrant/.ssh/authorized_keys /root/.ssh/authorized_keys
+    sudo chmod -R go-rwx /root/.ssh
+fi
